refactor(server): extract inline spec path helpers + RAII temp-dir cleanup (ARN-229)
Address PR review (Greptile P2s):
- Extract safe_inline_spec_path / safe_temp_dir_leaf and their tests into a new
  specs/paths.rs submodule; load_inline.rs drops from 716 to 621 lines.
- TempDirGuard removes the private per-request temp dir on Drop, so it is cleaned
  up on every exit path — success, a validation 400, or a write error — not just
  the happy path. Replaces the single post-load remove_dir_all.

Pure relocation plus RAII cleanup; no change to the validation behavior.
