+++ crates/temper-server/src/observe/specs/load_inline.rs
@@ -1,4 +1,4 @@
-use std::path::{Component, Path, PathBuf};
+use std::path::{Path, PathBuf};
 
 use axum::extract::State;
 use axum::http::HeaderMap;
@@ -17,6 +17,7 @@ use crate::authz::{DenialInput, record_authz_denial, security_context_from_heade
 use crate::state::{ServerState, TrajectoryEntry, TrajectorySource};
 
 use super::load_dir::handle_load_dir;
+use super::paths::{TempDirGuard, safe_inline_spec_path, safe_temp_dir_leaf};
 use super::types::{LoadDirRequest, LoadInlineRequest};
 
 /// POST /api/specs/load-inline -- load specs from inline content.
@@ -209,6 +210,9 @@ pub(crate) async fn handle_load_inline(
             format!("Failed to create temp dir: {e}"),
         )
     })?;
+    // Remove the private temp dir on every exit path — success, a validation 400,
+    // or a write error — so malformed requests can't accumulate abandoned dirs.
+    let _tmp_guard = TempDirGuard(tmp_dir.clone());
 
     let specs_root = resolve_inline_specs_root(&tmp_dir, &body.specs)?;
 
@@ -251,10 +255,7 @@ pub(crate) async fn handle_load_inline(
         merge: true,
     };
     let result = handle_load_dir(State(state.clone()), Json(dir_request)).await;
-
-    // The loader has read the specs into the registry; the per-request dir is no
-    // longer needed. Best-effort cleanup so unique temp dirs don't accumulate.
-    let _ = std::fs::remove_dir_all(&tmp_dir); // determinism-ok: HTTP handler removes its private temp dir after loading
+    // `_tmp_guard` removes the temp dir when this function returns (all paths).
 
     if result.is_ok()
         && let Some(ref cedar_text) = cedar_policies
@@ -403,59 +404,6 @@ fn resolve_inline_specs_root(
         .parent()
         .map(Path::to_path_buf)
         .unwrap_or_else(|| tmp_dir.to_path_buf()))
-}
-
-/// Resolve a caller-supplied inline-spec filename to a path guaranteed to stay
-/// under `tmp_dir`. The `specs` keys come from the request body, so a traversal
-/// (`../`) or absolute name must be rejected before any filesystem write escapes
-/// the private temp dir (ARN-229).
-fn safe_inline_spec_path(tmp_dir: &Path, filename: &str) -> Result<PathBuf, (StatusCode, String)> {
-    let rel = Path::new(filename);
-    if rel.as_os_str().is_empty() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            "inline spec path must not be empty".to_string(),
-        ));
-    }
-    // Every component must be a plain name: `..` is `ParentDir`, an absolute path
-    // starts with `RootDir`/`Prefix`, `.` is `CurDir` — all rejected, so the join
-    // can only ever land under `tmp_dir`. Nested relative paths (app subdirs) are
-    // allowed, since inline submissions legitimately carry them.
-    let mut safe = PathBuf::new();
-    for component in rel.components() {
-        match component {
-            Component::Normal(part) => safe.push(part),
-            _ => {
-                return Err((
-                    StatusCode::BAD_REQUEST,
-                    format!(
-                        "inline spec path '{filename}' must be a relative path with no '..', \
-                         '.', or absolute segments"
-                    ),
-                ));
-            }
-        }
-    }
-    Ok(tmp_dir.join(safe))
-}
-
-/// Resolve the per-request temp-dir leaf name to a direct child of `temp_root`.
-/// Unlike `safe_inline_spec_path` (which permits nested spec paths), the leaf must
-/// be exactly ONE `Component::Normal`: it embeds the caller-supplied tenant, and a
-/// tenant containing `/` would make only the final component uuid-suffixed, leaving
-/// a *predictable* intermediate dir a local attacker could pre-plant a symlink at
-/// (ARN-229). A single component keeps the whole dir unpredictable.
-fn safe_temp_dir_leaf(temp_root: &Path, leaf: &str) -> Result<PathBuf, (StatusCode, String)> {
-    let mut components = Path::new(leaf).components();
-    match (components.next(), components.next()) {
-        (Some(Component::Normal(part)), None) => Ok(temp_root.join(part)),
-        _ => Err((
-            StatusCode::BAD_REQUEST,
-            "invalid inline spec request: tenant must be a single safe path component \
-             (no '/', '..', or absolute segments)"
-                .to_string(),
-        )),
-    }
 }
 
 fn merge_inline_cedar_policy_text(existing: &str, incoming: &str) -> String {
@@ -618,50 +566,8 @@ mod tests {
 
     use super::{
         adr_candidate_paths, merge_inline_cedar_policy_text, namespace_to_app_candidates,
-        normalize_app_slug, safe_inline_spec_path, safe_temp_dir_leaf,
+        normalize_app_slug,
     };
-
-    #[test]
-    fn safe_inline_spec_path_rejects_traversal_and_absolute() {
-        // ARN-229: `specs` keys come from the request body. A traversal or absolute
-        // key must be rejected before it escapes the private temp dir.
-        let root = Path::new("/var/tmp/temper-inline-abc");
-        safe_inline_spec_path(root, "../../etc/cron.d/evil")
-            .expect_err("traversal key must be rejected");
-        safe_inline_spec_path(root, "/etc/passwd").expect_err("absolute key must be rejected");
-        safe_inline_spec_path(root, "..").expect_err("parent key must be rejected");
-        safe_inline_spec_path(root, "").expect_err("empty key must be rejected");
-        safe_inline_spec_path(root, "a/../../b").expect_err("mid-path traversal must be rejected");
-        // Legitimate relative spec paths (including nested app dirs) are allowed.
-        assert_eq!(
-            safe_inline_spec_path(root, "model.csdl.xml").expect("simple key allowed"),
-            root.join("model.csdl.xml")
-        );
-        assert_eq!(
-            safe_inline_spec_path(root, "app/order.ioa.toml").expect("nested key allowed"),
-            root.join("app/order.ioa.toml")
-        );
-    }
-
-    #[test]
-    fn safe_temp_dir_leaf_requires_single_component() {
-        // ARN-229: the tenant is interpolated into the temp-dir leaf name, which must
-        // stay a single unpredictable directory.
-        let temp_root = Path::new("/var/folders/tmp");
-        // Traversal tenant escapes the temp root.
-        safe_temp_dir_leaf(temp_root, "temper-inline-a/../../etc-01890000")
-            .expect_err("tenant traversal in the leaf name must be rejected");
-        // A tenant containing `/` (permitted by TenantId::new) would leave a
-        // predictable intermediate dir — reject the nested leaf.
-        safe_temp_dir_leaf(temp_root, "temper-inline-evil/x-01890000")
-            .expect_err("nested leaf (tenant with '/') must be rejected");
-        // A normal tenant yields a single child dir under the temp root.
-        assert_eq!(
-            safe_temp_dir_leaf(temp_root, "temper-inline-default-01890000")
-                .expect("normal tenant leaf allowed"),
-            temp_root.join("temper-inline-default-01890000")
-        );
-    }
 
     #[test]
     fn normalize_app_slug_kebab_cases_namespaces() {
